7. Modularize the code Refactor Refactor CourseList
