--- staff/gustavo-rojas/course-3/task-2/scheduler/kill_all.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-killall -s KILL node
-killall -s KILL "npm start"
