+++ staff/gustavo-rojas/course-3/task-2/scheduler/src/App.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from 'react';
 // import React from 'react';
 import './App.css';
-import { hasConflict,getCourseTerm,getCourseNumber,terms } from './utilities/times.js';
+// import { hasConflict,getCourseTerm,getCourseNumber,terms } from './utilities/times.js';
+import CourseList from './components/CourseList';
 const Banner = ({ title }) => (
   <h1>{ title }</h1>
 );
@@ -35,77 +36,12 @@ const addCourseTimes = course => ({
 
 
 
-const toggle = (x, lst) => (
-  lst.includes(x) ? lst.filter(y => y !== x) : [x, ...lst]
-);
-const Course = ({ course, selected, setSelected }) => {
-  const isSelected = selected.includes(course);
-  const isDisabled = !isSelected && hasConflict(course, selected);
-  const style = {
-    backgroundColor: isDisabled? 'lightgrey' : isSelected ? 'lightgreen' : 'white'
-  };
-  return (
-    <div className="card m-1 p-2" 
-      style={style}
-      onClick={isDisabled ? null : () =>  setSelected(toggle(course, selected))}>
-      <div className="card-body">
-        <div className="card-title">{ getCourseTerm(course) } CS { getCourseNumber(course) }</div>
-        <div className="card-text">{ course.title }</div>
-      </div>
-    </div>
-  );
-};
-
-const TermButton = ({term, setTerm, checked}) => (// We neeed change className for class  in this code class="btn btn-success m-1 p-2" htmlFor={term}>
-  <>
-    <input type="radio" id={term} className="btn-check" checked={checked} autoComplete="off"
-      onChange={() => setTerm(term)} />
-    <label className="btn btn-success m-1 p-2" htmlFor={term}>
-    { term }
-    </label>
-  </>
-);
-const TermSelector = ({term, setTerm}) => (
-  <div className="btn-group">
-  { 
-    Object.values(terms).map(value => (
-      <TermButton key={value} term={value} setTerm={setTerm} checked={value === term} />
-    ))
-  }
-  </div>
-);
-
-const CourseList = ({ courses }) => {
-  const [term, setTerm] = useState('Fall');
-  const [selected, setSelected] = useState([]);
-  const termCourses = Object.values(courses).filter(course => term === getCourseTerm(course));
-  
-  return (
-    <>
-      <TermSelector term={term} setTerm={setTerm} />
-      <div className="course-list">
-      { 
-        termCourses.map(course =>
-          <Course key={ course.id } course={ course }
-            selected={selected} setSelected={ setSelected } 
-          />) 
-      }
-      </div>
-    </>
-  );
-};
 
 
 
 
 
 
-
-
-
-
-
-//var setSchedule =[];
 
 const App = () => {
   const [schedule, setSchedule] = useState();
