+++ staff/gustavo-rojas/course-3/task-2/scheduler/src/components/CourseList.js
@@ -1,0 +1,65 @@
+import React, { useState } from 'react';
+import { hasConflict,getCourseTerm,getCourseNumber,terms } from '/mnt/c/ubiqum/ubiqum-bootcamp/staff/gustavo-rojas/course-3/task-2/scheduler/src/utilities/times.js';
+const CourseList = ({ courses }) => {
+    const [term, setTerm] = useState('Fall');
+    const [selected, setSelected] = useState([]);
+    const termCourses = Object.values(courses).filter(course => term === getCourseTerm(course));
+    
+    return (
+      <>
+        <TermSelector term={term} setTerm={setTerm} />
+        <div className="course-list">
+        { 
+          termCourses.map(course =>
+            <Course key={ course.id } course={ course }
+              selected={selected} setSelected={ setSelected } 
+            />) 
+        }
+        </div>
+      </>
+    );
+  };
+
+  const toggle = (x, lst) => (
+    lst.includes(x) ? lst.filter(y => y !== x) : [x, ...lst]
+  );
+
+  const TermButton = ({term, setTerm, checked}) => (// We neeed change className for class  in this code class="btn btn-success m-1 p-2" htmlFor={term}>
+  <>
+    <input type="radio" id={term} className="btn-check" checked={checked} autoComplete="off"
+      onChange={() => setTerm(term)} />
+    <label className="btn btn-success m-1 p-2" htmlFor={term}>
+    { term }
+    </label>
+  </>
+);
+
+const Course = ({ course, selected, setSelected }) => {
+  const isSelected = selected.includes(course);
+  const isDisabled = !isSelected && hasConflict(course, selected);
+  const style = {
+    backgroundColor: isDisabled? 'lightgrey' : isSelected ? 'lightgreen' : 'white'
+  };
+  return (
+    <div className="card m-1 p-2" 
+      style={style}
+      onClick={isDisabled ? null : () =>  setSelected(toggle(course, selected))}>
+      <div className="card-body">
+        <div className="card-title">{ getCourseTerm(course) } CS { getCourseNumber(course) }</div>
+        <div className="card-text">{ course.title }</div>
+      </div>
+    </div>
+  );
+};
+const TermSelector = ({term, setTerm}) => (
+    <div className="btn-group">
+    { 
+      Object.values(terms).map(value => (
+        <TermButton key={value} term={value} setTerm={setTerm} checked={value === term} />
+      ))
+    }
+    </div>
+  );
+  
+
+  export default CourseList;
